chore: small refactor to avoid dup code
