+++ crates/contract/tests/inprocess/mod.rs
@@ -1,2 +1,3 @@
 pub mod attestation_submission;
+pub mod common;
 pub mod update_participant_url;
